requirements.txt
ultralytics==8.3.31
numpy==1.26.3
opencv-python==4.10.0.84
matplotlib==3.9.2
pillow==10.2.0
ipykernel==6.29.5
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.17